requirements.txt
streamlit
graphviz
python-dotenv
pydantic
google-genai
